Java 8 modernization and cleanup in all packages
Removed final on static methods
Removed final on enum declarations
Remove unused imports
Removed methods that only call superclass method
Removed redundant access modifiers on enum constructor
Made as many fields possible final
Used lambda expressions where possible
Used Map::computeIfAbsent in place of verbose construct
Stopped creating redundant arrays in log statements
Removed unnecessary boxing and unboxing
Replaced .stream().forEach() with .forEach()
Used forEach loops in place of traditional iteration

Addresses #781 in part

+++ janusgraph-core/src/main/java/org/janusgraph/graphdb/database/serialize/attribute/ByteSerializer.java
@@ -24,12 +24,12 @@ public class ByteSerializer implements OrderPreservingSerializer<Byte> {
 
     @Override
     public Byte read(ScanBuffer buffer) {
-        return Byte.valueOf((byte)(buffer.getByte() + Byte.MIN_VALUE));
+        return (byte) (buffer.getByte() + Byte.MIN_VALUE);
     }
 
     @Override
     public void write(WriteBuffer out, Byte object) {
-        out.putByte((byte)(object.byteValue() - Byte.MIN_VALUE));
+        out.putByte((byte)(object - Byte.MIN_VALUE));
     }
 
     @Override
@@ -53,7 +53,7 @@ public class ByteSerializer implements OrderPreservingSerializer<Byte> {
             double d = ((Number)value).doubleValue();
             if (Double.isNaN(d) || Math.round(d)!=d) throw new IllegalArgumentException("Not a valid byte: " + value);
             long l = ((Number)value).longValue();
-            if (l>=Byte.MIN_VALUE && l<=Byte.MAX_VALUE) return Byte.valueOf((byte)l);
+            if (l>=Byte.MIN_VALUE && l<=Byte.MAX_VALUE) return (byte) l;
             else throw new IllegalArgumentException("Value too large for byte: " + value);
         } else if (value instanceof String) {
             return Byte.parseByte((String)value);
